Extract TimeAndScrubberGrid to common component
