+++ static/app/components/replays/replayController.tsx
@@ -6,13 +6,9 @@ import screenfull from 'screenfull';
 import {Button} from 'sentry/components/button';
 import ButtonBar from 'sentry/components/buttonBar';
 import {CompositeSelect} from 'sentry/components/compactSelect/composite';
-import ReplayTimeline from 'sentry/components/replays/breadcrumbs/replayTimeline';
-import {PlayerScrubber} from 'sentry/components/replays/player/scrubber';
-import useScrubberMouseTracking from 'sentry/components/replays/player/useScrubberMouseTracking';
 import {useReplayContext} from 'sentry/components/replays/replayContext';
-import {formatTime} from 'sentry/components/replays/utils';
+import TimeAndScrubberGrid from 'sentry/components/replays/timeAndScrubberGrid';
 import {
-  IconAdd,
   IconContract,
   IconExpand,
   IconNext,
@@ -21,7 +17,6 @@ import {
   IconPrevious,
   IconRewind10,
   IconSettings,
-  IconSubtract,
 } from 'sentry/icons';
 import {t} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
@@ -143,37 +138,6 @@ function ReplayOptionsMenu({speedOptions}: {speedOptions: number[]}) {
   );
 }
 
-function TimelineSizeBar() {
-  const {timelineScale, setTimelineScale, durationMs} = useReplayContext();
-  const maxScale = durationMs ? Math.ceil(durationMs / 60000) : 10;
-  return (
-    <ButtonBar>
-      <Button
-        size="xs"
-        title={t('Zoom out')}
-        icon={<IconSubtract />}
-        borderless
-        onClick={() => setTimelineScale(Math.max(timelineScale - 1, 1))}
-        aria-label={t('Zoom out')}
-        disabled={timelineScale === 1}
-      />
-      <span style={{padding: `0 ${space(0.5)}`}}>
-        {timelineScale}
-        {t('x')}
-      </span>
-      <Button
-        size="xs"
-        title={t('Zoom in')}
-        icon={<IconAdd />}
-        borderless
-        onClick={() => setTimelineScale(Math.min(timelineScale + 1, maxScale))}
-        aria-label={t('Zoom in')}
-        disabled={timelineScale === maxScale}
-      />
-    </ButtonBar>
-  );
-}
-
 function ReplayControls({
   toggleFullscreen,
   speedOptions = [0.1, 0.25, 0.5, 1, 2, 4, 8, 16],
@@ -183,7 +147,6 @@ function ReplayControls({
   const barRef = useRef<HTMLDivElement>(null);
   const [isCompact, setIsCompact] = useState(false);
   const isFullscreen = useIsFullscreen();
-  const {currentTime, startTimeOffsetMs, durationMs} = useReplayContext();
 
   // If the browser supports going fullscreen or not. iPhone Safari won't do
   // it. https://caniuse.com/fullscreen
@@ -211,34 +174,11 @@ function ReplayControls({
   });
   useLayoutEffect(() => updateIsCompact, [updateIsCompact]);
 
-  const elem = useRef<HTMLDivElement>(null);
-  const mouseTrackingProps = useScrubberMouseTracking({elem});
-
   return (
     <ButtonGrid ref={barRef} isCompact={isCompact}>
       <ReplayPlayPauseBar />
       <Container>
-        <TimeAndScrubberGrid id="replay-timeline-player" isCompact={isCompact}>
-          <Time style={{gridArea: 'currentTime'}}>
-            {formatTime(currentTime - startTimeOffsetMs)}
-          </Time>
-          <div style={{gridArea: 'timeline'}}>
-            <ReplayTimeline />
-          </div>
-          <div style={{gridArea: 'timelineSize', fontVariantNumeric: 'tabular-nums'}}>
-            <TimelineSizeBar />
-          </div>
-          <StyledScrubber
-            style={{gridArea: 'scrubber'}}
-            ref={elem}
-            {...mouseTrackingProps}
-          >
-            <PlayerScrubber showZoomIndicators />
-          </StyledScrubber>
-          <Time style={{gridArea: 'duration'}}>
-            {durationMs ? formatTime(durationMs) : '--:--'}
-          </Time>
-        </TimeAndScrubberGrid>
+        <TimeAndScrubberGrid isCompact={isCompact} showZoom />
       </Container>
       <ButtonBar gap={1}>
         <ReplayOptionsMenu speedOptions={speedOptions} />
@@ -269,36 +209,6 @@ const Container = styled('div')`
   flex-direction: column;
   flex: 1 1;
   justify-content: center;
-`;
-
-export const TimeAndScrubberGrid = styled('div')<{isCompact: boolean}>`
-  width: 100%;
-  display: grid;
-  grid-template-areas:
-    '. timeline timelineSize'
-    'currentTime scrubber duration';
-  grid-column-gap: ${space(1)};
-  grid-template-columns: max-content auto max-content;
-  align-items: center;
-  ${p =>
-    p.isCompact
-      ? `
-        order: -1;
-        min-width: 100%;
-        margin-top: -8px;
-      `
-      : ''}
-`;
-
-export const Time = styled('span')`
-  font-variant-numeric: tabular-nums;
-  padding: 0 ${space(1.5)};
-`;
-
-const StyledScrubber = styled('div')`
-  height: 32px;
-  display: flex;
-  align-items: center;
 `;
 
 export default ReplayControls;
